core: Install Chromium in GH action
Fix an issue in the "Deploy to GitHub Pages" action by adding a step to
explicitly install Chromium.

The prerendering step using Puppeteer was hanging indefinitely. There
may have been a problem with the automatic downloading of Chromium, or
it may have been an incompatability (as previously seen when developing
on Apple silicon). Either way, manually installing Chromium via apt
appears to resolve the issue.